feat: first version of dashboard
