+++ app/(app)/layout.tsx
@@ -7,9 +7,9 @@ export default function AppLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <div className="sm:grid grid-cols-[256px_1fr] min-h-screen">
+    <div className="sm:grid grid-cols-[256px_1fr] min-h-dvh">
       <Sidebar />
-      <main className="mx-auto w-full max-w-6xl">{children}</main>
+      <main className="h-[200vh] mx-auto w-full max-w-6xl">{children}</main>
     </div>
   );
 }
