+++ components/DashboardChart.tsx
@@ -1,0 +1,263 @@
+"use client";
+import { useState } from "react";
+import {
+  Area,
+  AreaChart,
+  CartesianGrid,
+  Label,
+  Pie,
+  PieChart,
+  XAxis,
+} from "recharts";
+import {
+  Card,
+  CardAction,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import {
+  type ChartConfig,
+  ChartContainer,
+  ChartLegend,
+  ChartLegendContent,
+  ChartTooltip,
+  ChartTooltipContent,
+} from "@/components/ui/chart";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+
+import { dashboardChartData } from "@/lib/data";
+
+const chartConfig = {
+  applied: {
+    label: "Applied",
+    color: "var(--chart-1)",
+  },
+  pending: {
+    label: "Pending",
+    color: "var(--chart-2)",
+  },
+  interviews: {
+    label: "Interviews",
+    color: "var(--chart-3)",
+  },
+  rejected: {
+    label: "Rejected",
+    color: "var(--chart-4)",
+  },
+  offered: {
+    label: "Offered",
+    color: "var(--chart-5)",
+  },
+} satisfies ChartConfig;
+
+export default function DashboardChart() {
+  const [timeRange, setTimeRange] = useState("90d");
+
+  const filteredData = dashboardChartData.filter((item) => {
+    const date = new Date(item.date);
+    const referenceDate = new Date("2026-05-10");
+    let daysToSubtract = 90;
+    if (timeRange === "30d") {
+      daysToSubtract = 30;
+    } else if (timeRange === "7d") {
+      daysToSubtract = 7;
+    }
+    const startDate = new Date(referenceDate);
+    startDate.setDate(startDate.getDate() - daysToSubtract);
+    return date >= startDate;
+  });
+
+  // 	let finalCharData = []
+  // 	Object.values(ApplicationStatus).map((status) => {
+  // 		let countCurrStatus = filteredData.filter((el) => el.status === status)
+  // 		finalCharData = [...finalCharData, {[status]: countCurrStatus.length}]
+  // 	})
+
+  return (
+    <Card className="pt-0 mb-14">
+      <CardHeader className="flex items-center gap-2 space-y-0 border-b py-5 sm:flex-row">
+        <div className="grid flex-1 gap-1">
+          <CardTitle>Application Journey</CardTitle>
+          <CardDescription>
+            Showing activity for the last{" "}
+            {timeRange === "90d"
+              ? "3 months"
+              : timeRange === "30d"
+                ? "1 month"
+                : "7 days"}
+          </CardDescription>
+        </div>
+        <Select value={timeRange} onValueChange={setTimeRange}>
+          <SelectTrigger
+            className="hidden w-[160px] rounded-lg sm:ml-auto sm:flex"
+            aria-label="Select a value"
+          >
+            <SelectValue placeholder="Last 3 months" />
+          </SelectTrigger>
+          <SelectContent className="rounded-xl">
+            <SelectItem value="90d" className="rounded-lg">
+              Last 3 months
+            </SelectItem>
+            <SelectItem value="30d" className="rounded-lg">
+              Last 30 days
+            </SelectItem>
+            <SelectItem value="7d" className="rounded-lg">
+              Last 7 days
+            </SelectItem>
+          </SelectContent>
+        </Select>
+      </CardHeader>
+      <CardContent className="px-2 pt-4 sm:px-6 sm:pt-6">
+        <ChartContainer
+          config={chartConfig}
+          className="aspect-auto h-[250px] w-full"
+        >
+          <AreaChart data={filteredData}>
+            <defs>
+              <linearGradient id="fillApplied" x1="0" y1="0" x2="0" y2="1">
+                <stop
+                  offset="5%"
+                  stopColor="var(--color-applied)"
+                  stopOpacity={0.8}
+                />
+                <stop
+                  offset="95%"
+                  stopColor="var(--color-applied)"
+                  stopOpacity={0.1}
+                />
+              </linearGradient>
+              <linearGradient id="fillPending" x1="0" y1="0" x2="0" y2="1">
+                <stop
+                  offset="5%"
+                  stopColor="var(--color-pending)"
+                  stopOpacity={0.8}
+                />
+                <stop
+                  offset="95%"
+                  stopColor="var(--color-pending)"
+                  stopOpacity={0.1}
+                />
+              </linearGradient>
+
+              <linearGradient id="fillInterviews" x1="0" y1="0" x2="0" y2="1">
+                <stop
+                  offset="5%"
+                  stopColor="var(--color-interviews)"
+                  stopOpacity={0.8}
+                />
+                <stop
+                  offset="95%"
+                  stopColor="var(--color-interviews)"
+                  stopOpacity={0.1}
+                />
+              </linearGradient>
+
+              <linearGradient id="fillRejected" x1="0" y1="0" x2="0" y2="1">
+                <stop
+                  offset="5%"
+                  stopColor="var(--color-rejected)"
+                  stopOpacity={0.8}
+                />
+                <stop
+                  offset="95%"
+                  stopColor="var(--color-rejected)"
+                  stopOpacity={0.1}
+                />
+              </linearGradient>
+
+              <linearGradient id="fillOffered" x1="0" y1="0" x2="0" y2="1">
+                <stop
+                  offset="5%"
+                  stopColor="var(--color-offered)"
+                  stopOpacity={0.8}
+                />
+                <stop
+                  offset="95%"
+                  stopColor="var(--color-offered)"
+                  stopOpacity={0.1}
+                />
+              </linearGradient>
+            </defs>
+            <CartesianGrid vertical={false} />
+            <XAxis
+              dataKey="date"
+              tickLine={false}
+              axisLine={false}
+              tickMargin={8}
+              minTickGap={32}
+              tickFormatter={(value) => {
+                const date = new Date(value);
+                return date.toLocaleDateString("en-US", {
+                  month: "short",
+                  day: "numeric",
+                });
+              }}
+            />
+            <ChartTooltip
+              cursor={false}
+              content={
+                <ChartTooltipContent
+                  labelFormatter={(value) => {
+                    return new Date(value).toLocaleDateString("en-US", {
+                      month: "short",
+                      day: "numeric",
+                    });
+                  }}
+                  indicator="dot"
+                />
+              }
+            />
+            <Area
+              dataKey="Applied"
+              type="natural"
+              fill="url(#fillApplied)"
+              stroke="var(--color-applied)"
+              stackId="a"
+            />
+            <Area
+              dataKey="Pending"
+              type="natural"
+              fill="url(#fillPending)"
+              stroke="var(--color-pending)"
+              stackId="a"
+            />
+
+            <Area
+              dataKey="Interviews"
+              type="natural"
+              fill="url(#fillInterviews)"
+              stroke="var(--color-interviews)"
+              stackId="a"
+            />
+
+            <Area
+              dataKey="Rejected"
+              type="natural"
+              fill="url(#fillRejected)"
+              stroke="var(--color-rejected)"
+              stackId="a"
+            />
+
+            <Area
+              dataKey="Offered"
+              type="natural"
+              fill="url(#fillOffered)"
+              stroke="var(--color-offered)"
+              stackId="a"
+            />
+            <ChartLegend content={<ChartLegendContent />} />
+          </AreaChart>
+        </ChartContainer>
+      </CardContent>
+    </Card>
+  );
+}
