+++ lib/data.ts
@@ -1,3 +1,5 @@
+import type { DashboardDataCards } from "@lib/types.ts";
+
 import {
   ChartColumnBig,
   CirclePlus,
@@ -16,4 +18,389 @@ export const sideLinks = [
   { href: "/applications", name: "Applications", icon: Folder },
   { href: "/new-entry", name: "New Entry", icon: CirclePlus },
   { href: "/analytics", name: "Analytics", icon: ChartColumnBig },
+];
+
+export const dashboardProfileDataCards: DashboardDataCards[] = [
+  {
+    title: "Applied to int.",
+    value: 18.7,
+    trendPercentage: 12,
+    trendDirection: "up",
+    trendLabel: "From last week",
+  },
+  {
+    title: "Int. to offer",
+    value: 12.5,
+    trendPercentage: 6,
+    trendDirection: "up",
+    trendLabel: "From last week",
+  },
+  {
+    title: "Response rate",
+    value: 64,
+    trendPercentage: 13,
+    trendDirection: "up",
+    trendLabel: "From last week",
+  },
+  {
+    title: "Average response",
+    value: 4.2,
+    trendPercentage: 17,
+    trendDirection: "down",
+    trendLabel: "From last week",
+  },
+];
+
+export const dashboardDataCards: DashboardDataCards[] = [
+  {
+    title: "Total Applications",
+    value: 128,
+    trendPercentage: 12,
+    trendDirection: "up",
+    trendLabel: "From last week",
+  },
+  {
+    title: "Interviews",
+    value: 24,
+    trendPercentage: 6,
+    trendDirection: "up",
+    trendLabel: "From last week",
+  },
+  {
+    title: "Offers",
+    value: 3,
+    trendPercentage: 13,
+    trendDirection: "up",
+    trendLabel: "From last week",
+  },
+  {
+    title: "Growth Rate",
+    value: 4.5,
+    trendPercentage: 17,
+    trendDirection: "down",
+    trendLabel: "From last week",
+  },
+];
+
+export const dashboardChartData = [
+  {
+    date: "2026-04-01",
+    Applied: 222,
+    Pending: 150,
+    Interviews: 180,
+    Rejected: 90,
+    Offered: 60,
+  },
+  {
+    date: "2026-04-02",
+    Applied: 97,
+    Pending: 180,
+    Interviews: 120,
+    Rejected: 70,
+    Offered: 40,
+  },
+  {
+    date: "2026-04-03",
+    Applied: 167,
+    Pending: 120,
+    Interviews: 140,
+    Rejected: 60,
+    Offered: 50,
+  },
+  {
+    date: "2026-04-04",
+    Applied: 242,
+    Pending: 260,
+    Interviews: 200,
+    Rejected: 110,
+    Offered: 80,
+  },
+  {
+    date: "2026-04-05",
+    Applied: 373,
+    Pending: 290,
+    Interviews: 250,
+    Rejected: 130,
+    Offered: 100,
+  },
+  {
+    date: "2026-04-06",
+    Applied: 301,
+    Pending: 340,
+    Interviews: 270,
+    Rejected: 120,
+    Offered: 90,
+  },
+  {
+    date: "2026-04-07",
+    Applied: 245,
+    Pending: 180,
+    Interviews: 190,
+    Rejected: 100,
+    Offered: 70,
+  },
+  {
+    date: "2026-04-08",
+    Applied: 409,
+    Pending: 320,
+    Interviews: 300,
+    Rejected: 150,
+    Offered: 120,
+  },
+  {
+    date: "2026-04-09",
+    Applied: 59,
+    Pending: 110,
+    Interviews: 90,
+    Rejected: 40,
+    Offered: 20,
+  },
+  {
+    date: "2026-04-10",
+    Applied: 261,
+    Pending: 190,
+    Interviews: 210,
+    Rejected: 100,
+    Offered: 75,
+  },
+  {
+    date: "2026-04-11",
+    Applied: 327,
+    Pending: 350,
+    Interviews: 280,
+    Rejected: 140,
+    Offered: 110,
+  },
+  {
+    date: "2026-04-12",
+    Applied: 292,
+    Pending: 210,
+    Interviews: 220,
+    Rejected: 110,
+    Offered: 85,
+  },
+  {
+    date: "2026-04-13",
+    Applied: 342,
+    Pending: 380,
+    Interviews: 300,
+    Rejected: 150,
+    Offered: 120,
+  },
+  {
+    date: "2026-04-14",
+    Applied: 137,
+    Pending: 220,
+    Interviews: 160,
+    Rejected: 80,
+    Offered: 50,
+  },
+  {
+    date: "2026-04-15",
+    Applied: 120,
+    Pending: 170,
+    Interviews: 150,
+    Rejected: 70,
+    Offered: 45,
+  },
+  {
+    date: "2026-04-16",
+    Applied: 138,
+    Pending: 190,
+    Interviews: 170,
+    Rejected: 85,
+    Offered: 55,
+  },
+  {
+    date: "2026-04-17",
+    Applied: 446,
+    Pending: 360,
+    Interviews: 320,
+    Rejected: 160,
+    Offered: 130,
+  },
+  {
+    date: "2026-04-18",
+    Applied: 364,
+    Pending: 410,
+    Interviews: 330,
+    Rejected: 170,
+    Offered: 140,
+  },
+  {
+    date: "2026-04-19",
+    Applied: 243,
+    Pending: 180,
+    Interviews: 200,
+    Rejected: 100,
+    Offered: 70,
+  },
+  {
+    date: "2026-04-20",
+    Applied: 89,
+    Pending: 150,
+    Interviews: 120,
+    Rejected: 60,
+    Offered: 35,
+  },
+  {
+    date: "2026-04-21",
+    Applied: 137,
+    Pending: 200,
+    Interviews: 180,
+    Rejected: 90,
+    Offered: 60,
+  },
+  {
+    date: "2026-04-22",
+    Applied: 224,
+    Pending: 170,
+    Interviews: 190,
+    Rejected: 95,
+    Offered: 70,
+  },
+  {
+    date: "2026-04-23",
+    Applied: 138,
+    Pending: 230,
+    Interviews: 200,
+    Rejected: 100,
+    Offered: 75,
+  },
+  {
+    date: "2026-04-24",
+    Applied: 387,
+    Pending: 290,
+    Interviews: 260,
+    Rejected: 130,
+    Offered: 100,
+  },
+  {
+    date: "2026-04-25",
+    Applied: 215,
+    Pending: 250,
+    Interviews: 220,
+    Rejected: 110,
+    Offered: 80,
+  },
+  {
+    date: "2026-04-26",
+    Applied: 75,
+    Pending: 130,
+    Interviews: 110,
+    Rejected: 50,
+    Offered: 30,
+  },
+  {
+    date: "2026-04-27",
+    Applied: 383,
+    Pending: 420,
+    Interviews: 340,
+    Rejected: 180,
+    Offered: 150,
+  },
+  {
+    date: "2026-04-28",
+    Applied: 122,
+    Pending: 180,
+    Interviews: 160,
+    Rejected: 80,
+    Offered: 55,
+  },
+  {
+    date: "2026-04-29",
+    Applied: 315,
+    Pending: 240,
+    Interviews: 230,
+    Rejected: 120,
+    Offered: 90,
+  },
+  {
+    date: "2026-04-30",
+    Applied: 454,
+    Pending: 380,
+    Interviews: 350,
+    Rejected: 190,
+    Offered: 160,
+  },
+  {
+    date: "2026-05-01",
+    Applied: 165,
+    Pending: 220,
+    Interviews: 180,
+    Rejected: 90,
+    Offered: 65,
+  },
+  {
+    date: "2026-05-02",
+    Applied: 293,
+    Pending: 310,
+    Interviews: 260,
+    Rejected: 140,
+    Offered: 110,
+  },
+  {
+    date: "2026-05-03",
+    Applied: 247,
+    Pending: 190,
+    Interviews: 210,
+    Rejected: 100,
+    Offered: 75,
+  },
+  {
+    date: "2026-05-04",
+    Applied: 385,
+    Pending: 420,
+    Interviews: 330,
+    Rejected: 170,
+    Offered: 140,
+  },
+  {
+    date: "2026-05-05",
+    Applied: 481,
+    Pending: 390,
+    Interviews: 360,
+    Rejected: 200,
+    Offered: 170,
+  },
+  {
+    date: "2026-05-06",
+    Applied: 498,
+    Pending: 520,
+    Interviews: 400,
+    Rejected: 220,
+    Offered: 180,
+  },
+  {
+    date: "2026-05-07",
+    Applied: 388,
+    Pending: 300,
+    Interviews: 290,
+    Rejected: 150,
+    Offered: 120,
+  },
+  {
+    date: "2026-05-08",
+    Applied: 149,
+    Pending: 210,
+    Interviews: 180,
+    Rejected: 90,
+    Offered: 60,
+  },
+  {
+    date: "2026-05-09",
+    Applied: 227,
+    Pending: 180,
+    Interviews: 190,
+    Rejected: 95,
+    Offered: 70,
+  },
+  {
+    date: "2026-05-10",
+    Applied: 293,
+    Pending: 330,
+    Interviews: 260,
+    Rejected: 140,
+    Offered: 110,
+  },
 ];
