+++ lib/types.ts
@@ -4,3 +4,11 @@ export interface SectionHeaderProps {
   cnTitle?: string;
   cnDescription?: string;
 }
+
+export interface DashboardDataCards {
+  title: string;
+  value: number;
+  trendPercentage: number;
+  trendDirection: "up" | "down";
+  trendLabel: string;
+}
